fix: apply module search paths restriction on worker and child process (#41118)
