+++ spec/fixtures/module/module-paths.js
@@ -1,0 +1,2 @@
+const Module = require('node:module');
+process.send(Module._nodeModulePaths(process.resourcesPath + '/test.js'));
