+++ spec/node-spec.ts
@@ -23,6 +23,12 @@ describe('node feature', () => {
         const [msg] = await message;
         expect(msg).to.equal('message');
       });
+
+      it('Has its module searth paths restricted', async () => {
+        const child = childProcess.fork(path.join(fixtures, 'module', 'module-paths.js'));
+        const [msg] = await once(child, 'message');
+        expect(msg.length).to.equal(2);
+      });
     });
   });
